+++ src/rtc/hacked-rtc-offer-options.ts
@@ -1,5 +1,0 @@
-export interface HackedRTCOfferOptions {
-    // FIXME @types/webrtc defines this interface to use numbers instead of booleans.
-    offerToReceiveAudio: boolean;
-    offerToReceiveVideo: boolean;
-}