+++ src/rtc/rtc-config.ts
@@ -1,8 +1,5 @@
-import { HackedRTCOfferOptions } from './hacked-rtc-offer-options';
-import { RTCAnswerOptions } from './rtc-answer-options';
-
 export interface RTCConfig extends RTCConfiguration {
-    defaultOfferOptions?: HackedRTCOfferOptions;
+    defaultOfferOptions?: RTCOfferOptions;
     defaultAnswerOptions?: RTCAnswerOptions;
     rtcpMuxPolicy?: 'require' | 'negotiate';
     bundlePolicy?: 'balanced' | 'max-compat' | 'max-bundle';